+++ src/ext/java/org/opentripplanner/ext/siri/TimetableHelper.java
@@ -19,7 +19,7 @@ import org.opentripplanner.model.Timetable;
 import org.opentripplanner.model.TimetableSnapshot;
 import org.opentripplanner.model.Trip;
 import org.opentripplanner.routing.RoutingService;
-import org.opentripplanner.routing.algorithm.raptor.transit.mappers.DateMapper;
+import org.opentripplanner.routing.algorithm.raptoradapter.transit.mappers.DateMapper;
 import org.opentripplanner.routing.graph.Graph;
 import org.opentripplanner.routing.trippattern.RealTimeState;
 import org.opentripplanner.routing.trippattern.TripTimes;
@@ -181,9 +181,7 @@ public class TimetableHelper {
                     } else if (recordedCall.getAimedDepartureTime() != null) {
                         realtimeDepartureTime = DateMapper.secondsSinceStartOfService(departureDate, recordedCall.getAimedDepartureTime(), zoneId);
                     }
-                    if (realtimeDepartureTime < realtimeArrivalTime) {
-                        realtimeDepartureTime = realtimeArrivalTime;
-                    }
+
                     int departureDelay = realtimeDepartureTime - departureTime;
 
                     newTimes.updateDepartureDelay(callCounter, departureDelay);
@@ -260,9 +258,6 @@ public class TimetableHelper {
 
                         if (realtimeArrivalTime == -1) {
                             realtimeArrivalTime = realtimeDepartureTime;
-                        }
-                        if (realtimeDepartureTime < realtimeArrivalTime) {
-                            realtimeDepartureTime = realtimeArrivalTime;
                         }
 
                         int arrivalDelay = realtimeArrivalTime - arrivalTime;